backend/backend_app/requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
docstring-parser==0.14.1
executing==1.2.0
msgpack==1.0.5
psycopg2-binary==2.9.6
py-algorand-sdk==2.1.2
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
pyteal==0.24.1
pytz==2023.3
semantic-version==2.10.0
sqlparse==0.4.4
tabulate==0.9.0
